update notebook with working prev/next
